feat: document auth headers

+++ ascella/src/database/queries/get_images.rs
@@ -6,11 +6,7 @@ pub async fn exec(owner: i32, amount: i32, skip: i32) -> Result<Vec<SimpleImages
     .await
     .query(
       //SQL INJECTION LMAOO this lang sucks!
-      format!(
-        "SELECT created, id,vanity FROM images WHERE owner = {} ORDER BY id DESC LIMIT {} OFFSET {}",
-        owner, amount, skip
-      )
-      .as_str(),
+      format!("SELECT created, id,vanity FROM images WHERE owner = {} ORDER BY id DESC LIMIT {} OFFSET {}", owner, amount, skip).as_str(),
       &[],
     )
     .await?
@@ -24,10 +20,7 @@ pub async fn exec(owner: i32, amount: i32, skip: i32) -> Result<Vec<SimpleImages
 pub async fn delete_all(owner: i32, date: i32) -> Result<u64> {
   let row = get_tokio_postgres()
     .await
-    .execute(
-      &format!("DELETE FROM images WHERE created < NOW() - INTERVAL '{} days' AND owner = $1", date,),
-      &[&owner],
-    )
+    .execute(&format!("DELETE FROM images WHERE created < NOW() - INTERVAL '{} days' AND owner = $1", date,), &[&owner])
     .await?;
 
   Ok(row)
